support running the integration tests in parallel
