+++ SpectraLogic.SpectraRioBrokerClient.Integration.Test/SpectraLogic.SpectraRioBrokerClient.Integration.Test.csproj
@@ -56,6 +56,7 @@
     <Compile Include="SpectraRioBrokerClientFixture.cs" />
     <Compile Include="SpectraRioBrokerClientIntegrationTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SpectraRioBrokerClientSeqIntegrationTests.cs" />
     <Compile Include="StringExtensions.cs" />
     <Compile Include="ValidationErrorComparer.cs" />
   </ItemGroup>
@@ -64,8 +65,8 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="TestFiles\F2.txt" />
-    <EmbeddedResource Include="TestFiles\F1.txt" />
+    <EmbeddedResource Include="TestFiles\F2_No_Size.txt" />
+    <EmbeddedResource Include="TestFiles\F1_No_Size.txt" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SpectraLogic.SpectraRioBrokerClient\SpectraLogic.SpectraRioBrokerClient.csproj">
