+++ SpectraLogic.SpectraRioBrokerClient.Integration.Test/SpectraRioBrokerClientIntegrationTests.cs
@@ -32,6 +32,7 @@ using SpectraLogic.SpectraRioBrokerClient.Utils;
 namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 {
     [TestFixture]
+    [Parallelizable(ParallelScope.All)]
     public class SpectraRioBrokerClientIntegrationTests
     {
         private const int MaxPollingAttempts = 10;
@@ -39,11 +40,22 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
         private readonly ILog _log = LogManager.GetLogger("SpectraRioBrokerClientIntegrationTest");
 
+        [OneTimeSetUp]
+        public void Setup()
+        {
+            SpectraRioBrokerClientFixture.SetupTestData();
+        }
+        
+        [OneTimeTearDown]
+        public void TearDown()
+        {
+            Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
+            Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
+        }
+        
         private void ArchiveNewFilesOnlyTest(IEnumerable<ArchiveFile> list1, IEnumerable<ArchiveFile> list2,
             string message)
         {
-            try
-            {
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, list1);
                 var archiveJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Archive(archiveRequest);
 
@@ -94,11 +106,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 {
                     Assert.AreEqual("Completed", file.Status);
                 }
-            }
-            finally
-            {
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-            }
         }
 
         [Test]
@@ -106,16 +113,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
         {
             var fileName1 = "Archive@And&Delete#With$Special Characters/" + Guid.NewGuid();
 
-            try
-            {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
-                /***********
-                 * ARCHIVE *
-                 ***********/
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri())
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 10)
                 });
 
                 var archiveJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Archive(archiveRequest);
@@ -147,12 +147,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-            }
-            finally
-            {
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
-            }
         }
 
         [Test]
@@ -163,16 +157,14 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 /***********
                  * ARCHIVE *
                  ***********/
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false),
-                    new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName2, "F2.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName2}}, false)
                 });
 
@@ -210,9 +202,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 var restoreRequest = new RestoreRequest(SpectraRioBrokerClientFixture.BrokerName, new List<RestoreFile>
                 {
                     new RestoreFile(fileName1,
-                        $"{SpectraRioBrokerClientFixture.RestoreTempDir}/F1_restore.txt".ToFileUri()),
+                        "F1_restore.txt".ToDevNullUri()),
                     new RestoreFile(fileName2,
-                        $"{SpectraRioBrokerClientFixture.RestoreTempDir}/F2_restore.txt".ToFileUri(),
+                        "F2_restore.txt".ToDevNullUri(),
                         new ByteRange(0, 10))
                 });
 
@@ -248,9 +240,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 var deleteF2Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName2);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF2Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
             }
         }
 
@@ -262,18 +251,16 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 /***********
                  * ARCHIVE *
                  ***********/
-                const string relationshipName = "ep_net_test";
+                const string relationshipName = "ArchiveAndRestoreWithRelationship";
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false,
                         new HashSet<string> {relationshipName}),
-                    new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName2, "F2.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName2}}, false,
                         new HashSet<string> {relationshipName})
                 });
@@ -329,7 +316,7 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 ***********/
 
                 var restoreList = relationship.Objects.Select(obj => new RestoreFile(obj.Name,
-                    $"{SpectraRioBrokerClientFixture.RestoreTempDir}/{obj.Name}".ToFileUri()));
+                    $"{obj.Name}".ToDevNullUri()));
 
                 var restoreRequest = new RestoreRequest(SpectraRioBrokerClientFixture.BrokerName, restoreList);
                 var restoreJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Restore(restoreRequest);
@@ -377,9 +364,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 var deleteF2Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName2);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF2Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
             }
         }
 
@@ -392,21 +376,19 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 ArchiveNewFilesOnlyTest(
                     new List<ArchiveFile>
                     {
-                        new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(),
+                        new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(),
                             14, new Dictionary<string, string> {{"fileName", fileName1}}, false),
-                        new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2.txt".ToFileUri(),
+                        new ArchiveFile(fileName2, "F2.txt".ToAtoZUri(),
                             14, new Dictionary<string, string> {{"fileName", fileName2}}, false)
                     },
                     new List<ArchiveFile>
                     {
-                        new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(),
+                        new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(),
                             14, new Dictionary<string, string> {{"fileName", fileName1}}, false),
-                        new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2.txt".ToFileUri(),
+                        new ArchiveFile(fileName2, "F2.txt".ToAtoZUri(),
                             14, new Dictionary<string, string> {{"fileName", fileName2}}, false)
                     },
                     "All files already exist. Nothing to archive.");
@@ -428,19 +410,17 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 ArchiveNewFilesOnlyTest(
                     new List<ArchiveFile>
                     {
-                        new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(),
+                        new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(),
                             14, new Dictionary<string, string> {{"fileName", fileName1}}, false)
                     },
                     new List<ArchiveFile>
                     {
-                        new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(),
+                        new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(),
                             14, new Dictionary<string, string> {{"fileName", fileName1}}, false),
-                        new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2.txt".ToFileUri(),
+                        new ArchiveFile(fileName2, "F2.txt".ToAtoZUri(),
                             14, new Dictionary<string, string> {{"fileName", fileName2}}, false)
                     },
                     "Archive job completed successfully");
@@ -461,7 +441,7 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
             {
-                new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/not_found.txt".ToFileUri(),
+                new ArchiveFile(fileName1, "not_found.txt".ToFileUri(),
                     14,
                     new Dictionary<string, string> {{"fileName", fileName1}}),
             }, failFast: false);
@@ -502,15 +482,13 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 /***********
                  * ARCHIVE *
                  ***********/
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri()),
-                    new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2.txt".ToFileUri())
+                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1_No_Size.txt".ToFileUri()),
+                    new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2_No_Size.txt".ToFileUri())
                 });
 
                 var archiveJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Archive(archiveRequest);
@@ -546,23 +524,16 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 var deleteF2Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName2);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF2Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
             }
         }
 
         [Test]
         public void CancelArchiveJob()
         {
-            try
-            {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var fileName1 = "CancelArchiveJob_" + Guid.NewGuid();
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -588,12 +559,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 //    Assert.AreEqual("Canceled", file.Status);
                 //}
             }
-            finally
-            {
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
-            }
-        }
 
         [Test]
         public void CancelRestoreJob()
@@ -602,11 +567,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -628,7 +591,7 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 var restoreRequest = new RestoreRequest(SpectraRioBrokerClientFixture.BrokerName, new List<RestoreFile>
                 {
                     new RestoreFile(fileName1,
-                        $"{SpectraRioBrokerClientFixture.RestoreTempDir}/F1_restore.txt".ToFileUri())
+                        "F1_restore.txt".ToDevNullUri())
                 });
 
                 var restoreJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Restore(restoreRequest);
@@ -656,22 +619,22 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
             }
         }
 
         [Test]
         public void DeleteBrokersTest()
         {
+            const string brokerName = "DeleteBrokersTest";
+            SpectraRioBrokerClientFixture.SpectraRioBrokerClient.CreateBroker(new CreateBrokerRequest(brokerName, SpectraRioBrokerClientFixture.AgentName, SpectraRioBrokerClientFixture.GetAgentConfig()));
+            Assert.DoesNotThrow(() => SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBroker(new GetBrokerRequest(brokerName)));
+            
             SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteBroker(
-                new DeleteBrokerRequest(SpectraRioBrokerClientFixture.BrokerName));
+                new DeleteBrokerRequest(brokerName));
             Assert.That(
                 () => SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBroker(
-                    new GetBrokerRequest(SpectraRioBrokerClientFixture.BrokerName)),
+                    new GetBrokerRequest(brokerName)),
                 Throws.Exception.TypeOf<BrokerNotFoundException>());
-            SpectraRioBrokerClientFixture.CreateBroker_1();
         }
 
         [Test]
@@ -681,11 +644,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -712,8 +673,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
             }
         }
 
@@ -724,11 +683,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -753,125 +710,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-            }
-        }
-
-        [Test]
-        public void GetBrokerObjectsTest()
-        {
-            var fileName1 = "a_" + "GetBrokerObjectsTest_" + Guid.NewGuid();
-            var fileName2 = "b_" + "GetBrokerObjectsTest_" + Guid.NewGuid();
-
-            try
-            {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
-                /***********
-                 * ARCHIVE *
-                 ***********/
-                const string relationshipName = "ep_net_test";
-                var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
-                {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
-                        new Dictionary<string, string> {{"fileName", fileName1}}, false,
-                        new HashSet<string> {relationshipName}),
-                    new ArchiveFile(fileName2, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F2.txt".ToFileUri(), 14,
-                        new Dictionary<string, string> {{"fileName", fileName2}}, false,
-                        new HashSet<string> {relationshipName})
-                });
-
-                var archiveJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Archive(archiveRequest);
-
-                var pollingAttempts = 0;
-                do
-                {
-                    archiveJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetJob(
-                        new GetJobRequest(archiveJob.JobId));
-                    _log.Debug(archiveJob.Status);
-                    Thread.Sleep(TimeSpan.FromSeconds(PollingInterval));
-                    pollingAttempts++;
-                } while (archiveJob.Status.Status == JobStatusEnum.ACTIVE && pollingAttempts < MaxPollingAttempts);
-
-                Assert.Less(pollingAttempts, MaxPollingAttempts);
-                Assert.AreEqual(JobStatusEnum.COMPLETED, archiveJob.Status.Status);
-
-                var brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName));
-
-                Assert.AreEqual(2, brokerObjects.Objects.Count);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName,
-                        ObjectsSortByEnum.NAME,
-                        SortOrderEnum.ASC));
-
-                Assert.AreEqual(2, brokerObjects.Objects.Count);
-                Assert.AreEqual(fileName1, brokerObjects.Objects.ElementAt(0).Name);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName,
-                        ObjectsSortByEnum.NAME,
-                        SortOrderEnum.DESC));
-
-                Assert.AreEqual(2, brokerObjects.Objects.Count);
-                Assert.AreEqual(fileName2, brokerObjects.Objects.ElementAt(0).Name);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName, perPage: 1));
-
-                Assert.AreEqual(1, brokerObjects.Objects.Count);
-                Assert.AreEqual(fileName2, brokerObjects.Objects.ElementAt(0).Name);
-                Assert.AreEqual(2, brokerObjects.Page.TotalPages);
-                Assert.AreEqual(1, brokerObjects.Page.PageSize);
-                Assert.AreEqual(0, brokerObjects.Page.Number);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName, filename: fileName2));
-
-                Assert.AreEqual(1, brokerObjects.Objects.Count);
-                Assert.AreEqual(fileName2, brokerObjects.Objects.ElementAt(0).Name);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName, prefix: "a"));
-
-                Assert.AreEqual(1, brokerObjects.Objects.Count);
-                Assert.AreEqual(fileName1, brokerObjects.Objects.ElementAt(0).Name);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName, prefix: "a",
-                        filename: "not_found"));
-
-                Assert.AreEqual(0, brokerObjects.Objects.Count);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName,
-                        metadata: new List<KeyValuePair<string, string>>
-                        {
-                            new KeyValuePair<string, string>("fileName", fileName2)
-                        }));
-
-                Assert.AreEqual(1, brokerObjects.Objects.Count);
-                Assert.AreEqual(fileName2, brokerObjects.Objects.ElementAt(0).Name);
-
-                brokerObjects = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokerObjects(
-                    new GetBrokerObjectsRequest(SpectraRioBrokerClientFixture.BrokerName,
-                        relationships: new List<string>
-                        {
-                            "ep_net_test"
-                        }));
-
-                Assert.AreEqual(2, brokerObjects.Objects.Count);
-            }
-            finally
-            {
-                var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
-                var deleteF2Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName2);
-                SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-                SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF2Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
             }
         }
 
@@ -882,11 +720,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -918,8 +754,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
             }
         }
 
@@ -930,11 +764,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -966,79 +798,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-            }
-        }
-
-        [Test]
-        public void GetBrokersTest()
-        {
-            var brokers = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetBrokers(new GetBrokersRequest());
-            Assert.AreEqual(2, brokers.BrokerList.Count);
-        }
-
-        [Test]
-        public void GetJobsTest()
-        {
-            var fileName1 = Guid.NewGuid().ToString();
-            try
-            {
-                var archiveJobsBefore = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetJobs(
-                    new GetJobsRequest(
-                        jobTypes: new List<JobTypeEnum> {JobTypeEnum.ARCHIVE},
-                        jobStatuses: new List<JobStatusEnum> {JobStatusEnum.COMPLETED})
-                ).JobsList.Count;
-                var restoreJobsBefore = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetJobs(
-                    new GetJobsRequest(jobTypes: new List<JobTypeEnum> {JobTypeEnum.RESTORE})
-                ).JobsList.Count;
-
-                SpectraRioBrokerClientFixture.SetupTestData();
-                var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
-                {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
-                        new Dictionary<string, string> {{"fileName", fileName1}}, false)
-                });
-
-                var archiveJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Archive(archiveRequest);
-
-                var pollingAttempts = 0;
-                do
-                {
-                    archiveJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetJob(
-                        new GetJobRequest(archiveJob.JobId));
-                    _log.Debug(archiveJob.Status);
-                    Thread.Sleep(TimeSpan.FromSeconds(PollingInterval));
-                    pollingAttempts++;
-                } while (archiveJob.Status.Status == JobStatusEnum.ACTIVE && pollingAttempts < MaxPollingAttempts);
-
-                Assert.Less(pollingAttempts, MaxPollingAttempts);
-                Assert.AreEqual(JobStatusEnum.COMPLETED, archiveJob.Status.Status);
-
-                var jobs = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetJobs(
-                    new GetJobsRequest(
-                        jobTypes: new List<JobTypeEnum> {JobTypeEnum.ARCHIVE},
-                        jobStatuses: new List<JobStatusEnum> {JobStatusEnum.COMPLETED}));
-                Assert.AreEqual(archiveJobsBefore + 1, jobs.JobsList.Count);
-
-                jobs = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetJobs(
-                    new GetJobsRequest(jobTypes: new List<JobTypeEnum> {JobTypeEnum.RESTORE}));
-                Assert.AreEqual(restoreJobsBefore, jobs.JobsList.Count);
-
-                jobs = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.GetJobs(
-                    new GetJobsRequest(
-                        jobTypes: new List<JobTypeEnum> {JobTypeEnum.ARCHIVE},
-                        jobStatuses: new List<JobStatusEnum> {JobStatusEnum.COMPLETED},
-                        sortBy: JobsSortByEnum.CREATION_DATE,
-                        sortOrder: SortOrderEnum.DESC));
-                Assert.AreEqual(archiveJob.JobId, jobs.JobsList[0].JobId);
-            }
-            finally
-            {
-                var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
-                SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
             }
         }
 
@@ -1078,11 +837,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -1103,7 +860,7 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
                 archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName2, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -1126,7 +883,7 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 var restoreRequest = new RestoreRequest("*", new List<RestoreFile>
                 {
                     new RestoreFile(fileName1,
-                        $"{SpectraRioBrokerClientFixture.RestoreTempDir}/F1_restore.txt".ToFileUri())
+                        "F1_restore.txt".ToDevNullUri())
                 }, true);
 
                 var restoreJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Restore(restoreRequest);
@@ -1162,9 +919,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
                 deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName2, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
             }
         }
 
@@ -1175,11 +929,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -1235,9 +987,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
             }
         }
 
@@ -1248,11 +997,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -1274,7 +1021,7 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
                 var restoreRequest = new RestoreRequest(SpectraRioBrokerClientFixture.BrokerName, new List<RestoreFile>
                 {
                     new RestoreFile(fileName1,
-                        $"{SpectraRioBrokerClientFixture.RestoreTempDir}/F1_restore.txt".ToFileUri())
+                        "F1_restore.txt".ToDevNullUri())
                 });
 
                 var restoreJob = SpectraRioBrokerClientFixture.SpectraRioBrokerClient.Restore(restoreRequest);
@@ -1331,9 +1078,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest(SpectraRioBrokerClientFixture.BrokerName, fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
-                Directory.Delete(SpectraRioBrokerClientFixture.RestoreTempDir, true);
             }
         }
 
@@ -1344,11 +1088,9 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
 
             try
             {
-                SpectraRioBrokerClientFixture.SetupTestData();
-
                 var archiveRequest = new ArchiveRequest(SpectraRioBrokerClientFixture.BrokerName, new List<ArchiveFile>
                 {
-                    new ArchiveFile(fileName1, $"{SpectraRioBrokerClientFixture.ArchiveTempDir}/F1.txt".ToFileUri(), 14,
+                    new ArchiveFile(fileName1, "F1.txt".ToAtoZUri(), 14,
                         new Dictionary<string, string> {{"fileName", fileName1}}, false)
                 });
 
@@ -1371,8 +1113,6 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
             {
                 var deleteF1Request = new DeleteFileRequest("*", fileName1);
                 SpectraRioBrokerClientFixture.SpectraRioBrokerClient.DeleteFile(deleteF1Request);
-
-                Directory.Delete(SpectraRioBrokerClientFixture.ArchiveTempDir, true);
             }
         }
