+++ SpectraLogic.SpectraRioBrokerClient.Integration.Test/StringExtensions.cs
@@ -22,9 +22,18 @@ namespace SpectraLogic.SpectraRioBrokerClient.Integration.Test
     {
         #region Public Methods
 
-        public static Uri ToFileUri(this string str) => $"file://{str.Replace("\\", "/").Replace(" ", "%20")}".ToUri();
+        private static string EscapeUri(this string str) =>
+            str.Replace("\\", "/").Replace(" ", "%20");
+
+        public static Uri ToFileUri(this string str) => $"file://{str.EscapeUri()}".ToUri();
 
         public static Uri ToHttpsUri(this string str) => $"https://{str}".ToUri();
+
+        public static Uri ToAtoZUri(this string str) =>
+            $"aToZSequence://{str.EscapeUri()}".ToUri();
+
+        public static Uri ToDevNullUri(this string str) =>
+            $"null:///{str.EscapeUri()}".ToUri();
 
         #endregion Public Methods
     }
